vicinity/requirements.txt
decorator==4.0.6
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
ipython==4.1.1
ipython-genutils==0.1.0
Markdown==2.6.5
path.py==8.1.2
pickleshare==0.6
psycopg2==2.6.1
pyreadline==2.1
simplegeneric==0.8.1
traitlets==4.1.0
wheel==0.24.0
